added collision boxes and kick animations
