--- Assets/Evan-Scripts/FollowCamera.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 40320ceff9c7c7c4eb5ba663bb71720b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
