--- Assets/Evan-Scripts/LookAtCamera.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0462f93dcca95094387436bb10efe4b5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
