--- Assets/Evan-Scripts/MoveScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b37c30b0beac21740969e4caf931246d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
