--- Assets/Evan-Scripts/PlayerMove.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1485f1606afc1a3479757da10b6cecad
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
