--- Assets/Evan-Scripts/new character anim.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 17ce66a1a09744146ae754e40cd395bd
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
