+++ Assets/ourAnims/character anims.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6811291a64b91fc4cb45450791c78dff
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
